+++ qemu/util/oslib-win32.c
@@ -94,7 +94,7 @@ void qemu_anon_ram_free(void *ptr, size_t size)
     }
 }
 
-size_t getpagesize(void)
+int getpagesize(void)
 {
     SYSTEM_INFO system_info;
